% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_group.R
\name{dg_bar}
\alias{dg_bar}
\title{Calculate value of truncated normal density on nonnegative reals up to multiplicative constant}
\usage{
dg_bar(n, s, t, sig)
}
\arguments{
\item{n}{The dimension of original space}

\item{s}{The element of Rn / O(n) at which the value of the truncated normal
density is calculated up to a multiplicative constant (the proposed state)}

\item{t}{The element of Rn / O(n) at which the mode of the truncated normal
density occurs (the current state)}

\item{sig}{The scale parameter of the truncated normal}
}
\value{
value of trunctated normal density at t up to multiplicative constant
}
\description{
Calculate value of truncated normal density on nonnegative reals up to multiplicative constant
}
\examples{
sample <- dg_bar(s, t, sig);
}
