% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_group.R
\name{rg_bar}
\alias{rg_bar}
\title{Sample value of truncated normal density on nonnegative reals}
\usage{
rg_bar(s, t, sig)
}
\arguments{
\item{s}{The element of Rn / O(n) at which the value of the truncated normal
density is calculated up to a multiplicative constant (the proposed state)}

\item{t}{The element of Rn / O(n) at which the mode of the truncated normal
density occurs (the current state)}

\item{sig}{The scale parameter of the truncated normal}
}
\value{
value of trunctated normal density at t up to multiplicative constant
}
\description{
Sample value of truncated normal density on nonnegative reals
}
\examples{
sample <- rg_bar(s, t, sig);
}
