% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_group.R
\name{sample_qmh}
\alias{sample_qmh}
\title{Quotient Metropolis Hastings on Rn / O(n)}
\usage{
sample_qmh(f_bar)
}
\arguments{
\item{f_bar}{The induced density on the quotient}
}
\value{
Samples from original distribution
}
\description{
Sample from distribution using Quotient Metropolis-Hastings on Rn / O(n)
}
\examples{
samples <- sample_qmh(f_bar);
}
